requirements.txt
aniso8601==9.0.1
blinker==1.6.2
click==8.1.7
colorama==0.4.6
flask==2.3.3
Flask-Cors==4.0.0
Flask-RESTful==0.3.10
flask-sqlalchemy==3.0.5
greenlet==2.0.2
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
PyMySQL==1.1.0
pytz==2023.3.post1
six==1.16.0
SQLAlchemy==2.0.20
typing-extensions==4.7.1
werkzeug==2.3.7
zipp==3.16.2
PyJwt
